feat(fase-05): 5.11c explicit lifecycle + 5.19 B+tree batch delete
## 5.11c — Explicit connection state machine
- New lifecycle.rs: ConnectionPhase enum (Connected→Auth→Idle→Executing→Closing)
  separate from ConnectionState (SQL session state)
- LifecycleTimeouts: fixed 10s auth timeout; wait_timeout vs interactive_timeout
  on idle reads based on CLIENT_INTERACTIVE capability flag
- net_write_timeout applied to all packet writes; TCP_NODELAY + SO_KEEPALIVE on
  accepted sockets via socket2 helper
- COM_RESET_CONNECTION preserves interactive classification across session reset
- 3 live-socket integration tests (auth timeout, idle timeout, interactive reset)
- 11 wire assertions covering timeout vars, validation, idle close, interactive

## 5.19 — B+tree batch delete (sorted single-pass)
- BTree::delete_many_in: accepts pre-sorted encoded keys, removes all in one
  left-to-right O(N) pass via batch_delete_subtree/leaf/internal helpers
- collect_delete_key_buckets: groups exact encoded keys per index with NULL-skip,
  partial-index predicate gating, composite and FK key encoding
- delete_many_from_indexes: replaces per-row delete_from_indexes for DELETE and
  UPDATE old-key removal; root persisted once per affected index batch
- executor/delete.rs and update.rs updated to batch path
- Local bench 50K rows: delete_where 1.08M r/s (+48% vs MySQL 653K),
  delete bulk 7.6M r/s (7.6× MySQL); update gap documented as 5.20 target

## Backlog additions (progreso.md)
- 5.20: UPDATE index-skip for non-indexed columns (12× gap at 50K rows)
- 5.21: INSERT WAL group-commit throughput (3.4× gap at 50K rows)

## Docs + wire test
- docs-site/src/internals/btree.md: delete_many_in API + algorithm + callout
- docs-site/src/internals/architecture.md: lifecycle state machine section
- docs-site/src/internals/executor.md: batch delete path in bloom table
- wire-test.py: 8 new 5.19 assertions; header updated to 5.11c/5.19/5.19a
- 212/212 wire assertions passing

+++ crates/axiomdb-network/src/mysql/prepared.rs
@@ -967,8 +967,8 @@ mod tests {
         p.extend_from_slice(&1u32.to_le_bytes()); // iteration_count
         p.push(0x00); // null bitmap (neither is null at wire level)
         p.push(1); // new_params_bound_flag
-        p.extend_from_slice(&(param0_type as u16).to_le_bytes());
-        p.extend_from_slice(&(param1_type as u16).to_le_bytes());
+        p.extend_from_slice(&param0_type.to_le_bytes());
+        p.extend_from_slice(&param1_type.to_le_bytes());
         // param0 has long data — no inline bytes for it
         // param1: lenenc + value
         p.push(param1_value.len() as u8);
